misc/codegen/pyproject.toml
[project]
name = "codegen"
version = "1.0.0"
description = "Scripts for generating Device Path Nodes related code for EFI Boot Editor"

requires-python = ">=3.11"

dependencies = [
    "beautifulsoup4",
    "jinja2",
    "pyyaml",
]

[project.optional-dependencies]

dev = [
    "mypy",
    "pip-tools",
    "ruff",
    "types-pyyaml",
    "types-beautifulsoup4",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
select = [
    "B",
    "C",
    "E",
    "F",
    "I",
    "W",
]

line-length = 120

[tool.black]
line-length = 120

[tool.mypy]
strict = true
check_untyped_defs = true
disallow_any_generics = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

plugins = []
